Removed 2D Manual Object Tracking namespace with everything in it.
## Reason
This was a failed attempt to replace Unity's built-in physics engine wtih only features I wanted, while developing SpaceNations.
It kind of did what I wanted it to: it was predictable, constant, and stable. But performance-wise it wasn't good enough -- it was still better than Unity's Box2D though.
Due to the fact that Unity guys are now exposing Box2D internals, this library is beyond redundant.
That being said, after that disappointment with SpaceNations project, Unity will not be my first choice for a performance-critical 2D game anymore.