Removed UniqueIdGenerator class.
## Reason
This was one of the hardest calls.
The class is simple and robust, it doesn exactly waht you ask for without any chance of corruption. BUT, you don't have control over how it is done.
These type of classes usually are used as a part of some bigger structures, in my case, this class was introduced due to the fact tat I was planning to develop a save-game system and needed a way to identify objects that will be saved.
The save-game system never got implemented and this class was left hanging on.
So in essence, this class should not exist on its own, but rather exist as a part of the system i belongs.